делает из карты компонент
