--- src/components/card/card.jsx
@@ -1,40 +1,0 @@
-import {React, PureComponent} from "react";
-import PropTypes from "prop-types";
-import leaflet from "leaflet";
-
-const Card = (props) => {
-  const {hotels} = props;
-
-  const city = [52.38333, 4.9];
-
-  const icon = leaflet.icon({
-    iconUrl: `../../../public/img/pin-1.svg`,
-    iconSize: [30, 30]
-  });
-
-  const zoom = 12;
-  const map = leaflet.map(`map`, {
-    center: city,
-    zoom: zoom,
-    zoomControl: false,
-    marker: true
-  });
-
-  map.setView(city, zoom);
-
-  leaflet
-  .tileLayer(`https://{s}.basemaps.cartocdn.com/rastertiles/voyager/{z}/{x}/{y}{r}.png`, { 
-    attribution: `&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors &copy; <a href="https://carto.com/attributions">CARTO</a>`
-  })
-  .addTo(map);
-
-  const offerCords = [52.3709553943508, 4.89309666406198];
-  leaflet
-  .marker(offerCords, {icon})
-  .addTo(map);
-
-  return <div id="map" style={{height: `100%`}}></div>
-};
-
-
-export default Card;
