+++ src/components/hotels-page/hotels-page.jsx
@@ -1,7 +1,7 @@
 import React, {PureComponent} from "react";
 import PropTypes from "prop-types";
 
-import Card from "../card/card.jsx";
+import Map from "../map/map.jsx";
 import HotelCard from "../hotel-card/hotel-card.jsx";
 
 class HotelsPage extends PureComponent {
@@ -24,6 +24,8 @@ class HotelsPage extends PureComponent {
   render() {
     const {hotels} = this.props;
 
+  const map = <Map />;
+
     const listItems = hotels.map((item, i) => (
       <HotelCard
         key={i}
@@ -32,8 +34,6 @@ class HotelsPage extends PureComponent {
         onClick={this._onClick}
       />
     ));
-
-    const card = <Card hotels={hotels} />
 
     return <React.Fragment>
       <div style={{display: `none`}}>
@@ -134,9 +134,7 @@ class HotelsPage extends PureComponent {
               </div>
             </section>
             <div className="cities__right-section">
-              <section className="cities__map map">
-                {card}
-              </section>
+              {map}
             </div>
           </div>
         </div>
